+++ packages/jest-transform/src/ScriptTransformer.ts
@@ -773,18 +773,18 @@ class ScriptTransformer {
         },
       },
     );
-    const module: ModuleType = await requireOrImportModule(
-      moduleName,
-      applyInteropRequireDefault,
-    );
-
-    if (!callback) {
-      revertHook();
-
-      return module;
-    }
-
     try {
+      const module: ModuleType = await requireOrImportModule(
+        moduleName,
+        applyInteropRequireDefault,
+      );
+
+      if (!callback) {
+        revertHook();
+
+        return module;
+      }
+
       const cbResult = callback(module);
 
       if (isPromise(cbResult)) {
@@ -792,11 +792,11 @@ class ScriptTransformer {
           () => module,
         );
       }
+
+      return module;
     } finally {
       revertHook();
     }
-
-    return module;
   }
 
   shouldTransform(filename: Config.Path): boolean {